feat(spinner): add Output method to expose overwriting the output
This only applies under Accessible mode (which seems to be the original intention), leaving the tea program writing to Stderr.

The reason for this is under some environments it may be preferable to have Accessible write to stderr - Our usecase is that the spinner produces some json output that is piped to `jq` in some automated environments, but also availble as a CLI for users.