dynamic default values in input fields
Default values are a special type of placeholder that is treated as the field's actual value when the value would otherwise be empty. The existing default value mechanism, the field's "initial value" is fully static in its current implementation. Although it may be possible to make that mechanism more dynamic in-place, that would have been a much more invasive change. Treating default values more like the placeholder value is more consistent with the rest of the dynamic forms API and also yields a slightly better default user experience.